+++ Example/ExampleApp/AppDelegate.swift
@@ -20,12 +20,15 @@ import UIKit
 
 @UIApplicationMain
 final class AppDelegate: UIResponder, UIApplicationDelegate {
-
-    var window: UIWindow?
+    let window = UIWindow()
 
     func application(_ application: UIApplication,
                      // swiftlint:disable:next discouraged_optional_collection
                      didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil) -> Bool {
-        true
+        let viewController = CompanyViewController()
+        let navigationController = UINavigationController(rootViewController: viewController)
+        self.window.rootViewController = navigationController
+        self.window.makeKeyAndVisible()
+        return true
     }
 }